int32 -> int64, outputs's display improved

+++ src/Netlist/template.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int print(unsigned int a){
-	printf("%u\n", a);
+int print(unsigned long int a){
+	printf("%lu\n", a);
 }
